split register files/functions
