+++ Register.cpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <iostream>
+#include <fstream>
+#include <string>
+
+
+class Register {
+private:
+
+public:
+	int Loc;			// RF Location
+	int val;			// Value of Location
+
+};
