+++ RegisterAllocationTable.cpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <iostream>
+#include <fstream>
+#include <string>
+
+class RegisterAllocationTable {
+private:
+
+public:
+	int callVal();
+};
+
+int toInt(std::string str);
