+++ RegisterFile.cpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <iostream>
+#include <fstream>
+#include <string>
+
+class RegisterFile {
+private:
+
+public:
+
+};
