+++ ReservationStation.cpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <iostream>
+#include <fstream>
+#include <string>
+
+class ReservationStation {
+private:
+
+public:
+	void Dispatch();
+};
